model/deblurring-master/requirements.txt
alabaster==0.7.10
awscli==1.11.174
Babel==2.5.1
bleach==1.5.0
botocore==1.7.32
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.7
coverage==4.4.1
docutils==0.14
enum34==1.1.6
flake8==3.4.1
h5py==2.7.1
html5lib==0.9999999
idna==2.6
imagesize==0.7.1
imutils==0.4.3
Jinja2==2.9.6
jmespath==0.9.3
Keras==2.1.1
Markdown==2.6.9
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
Pillow==4.3.0
protobuf==3.4.0
pyasn1==0.3.7
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.11
scipy==1.0.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.4
sphinxcontrib-websupport==1.0.1
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc2
tqdm==4.19.4
urllib3==1.22
Werkzeug==0.12.2
